requirements.txt
Django>=3.1
django-money
psycopg2-binary
psycopg2
django-rest-framework
djangorestframework 
gunicorn 
PyJWT
pytz
